+++ docs/source/changelog.rst
@@ -28,6 +28,33 @@ Changelog
 .. - UNSOLVED (:issue:`397`) extras failed
 
 
+v0.32.0 / 2025-MM-DD (Unreleased)
+--------------------
+
+Breaking Changes
+++++++++++++++++
+
+New Features
+++++++++++++
+- UNMERGED (:pr:`470`) Psi4, CFOUR, NWChem - Demonstrated running energies and gradients with additional
+  point charges. @loriab
+
+Enhancements
+++++++++++++
+- UNMERGED (:pr:`470`) CFOUR - Adapted gradient parsing to handle point charges. Adapted options parsing
+  to allow `%<something>*` sections. @loriab
+- UNMERGED (:pr:`470`) NWChem - Adapted options parsing for nested arrays, particularly for `bq` section.
+  Use "data", e.g., `bq__data` to fill in keywordless lines into blocks.
+- UNMERGED (:pr:`470`) NWChem - Broaden allowing autoz, noautoz, center, nocenter, autosym, noautosym
+  keywords for molecule.
+
+Bug Fixes
++++++++++
+
+Misc.
++++++
+
+
 v0.31.0 / 2025-01-17
 --------------------